+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs-Carbon/General/Axes/LabelHidden.jsx
@@ -3,12 +3,13 @@ import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/terra-graphs-src/components/Graph.module.scss';
 import ExampleGraphContainer from '@cerner/terra-graphs-docs/lib/terra-dev-site/ExampleGraphContainer/ExampleGraphContainer';
-import lineDefault from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineDefault';
+import getGraphConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/General/generalDefaultLabelHidden';
 
 /*
 Please refer to the documentation below to see the graphConfig and data objects
 */
-const graphConfig = utils.deepClone(lineDefault('#labelHidden'));
+
+const graphConfig = utils.deepClone(getGraphConfig('#labelHidden'));
 graphConfig.showLabel = false;
 
 const LabelHiddenGeneralExample = () => {