+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs-Carbon/General/Axes/StaticXAxisTicks.jsx
@@ -3,24 +3,13 @@ import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/terra-graphs-src/components/Graph.module.scss';
 import ExampleGraphContainer from '@cerner/terra-graphs-docs/lib/terra-dev-site/ExampleGraphContainer/ExampleGraphContainer';
-import lineTimeseries from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineTimeseries';
-
-const tickValues = [
-  new Date(2016, 0, 1, 1, 0).toISOString(),
-  new Date(2016, 0, 1, 5, 0).toISOString(),
-  new Date(2016, 0, 1, 10, 0).toISOString(),
-  new Date(2016, 0, 1, 15, 0).toISOString(),
-  new Date(2016, 0, 1, 20, 0).toISOString(),
-];
+import getGraphConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/General/timeseriesConfig';
 
 /*
 Please refer to the documentation below to see the graphConfig and data objects
 */
-const graphConfig = utils.deepClone(lineTimeseries('#staticXAxisTicks'));
-graphConfig.axis.x.ticks = {
-  values: tickValues,
-  format: '%H:%M:%S',
-};
+
+const graphConfig = utils.deepClone(getGraphConfig('#staticXAxisTicks'));
 
 const StaticXAxisTicksGeneralExample = () => {
   React.useEffect(() => {