+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs-Carbon/General/Legend/LegendHidden.jsx
@@ -4,22 +4,20 @@ import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/terra-graphs-src/components/Graph.module.scss';
 import '@cerner/terra-graphs-docs/lib/terra-graphs-src/components/Line/LineGraph.module.scss';
 import ExampleGraphContainer from '@cerner/terra-graphs-docs/lib/terra-dev-site/ExampleGraphContainer/ExampleGraphContainer';
-import lineDefault from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineDefault';
-import exampleData from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/simpleLineData';
+import getGraphConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/General/generalDefaultLegendHidden';
+import exampleData from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/General/dataset1';
 
 /*
 Please refer to the documentation below to see the graphConfig and data objects
 */
-const graphConfig = utils.deepClone(lineDefault('#legendHidden'));
-graphConfig.showLegend = false;
+
+const graphConfig = utils.deepClone(getGraphConfig('#legendHidden'));
 const dataset = utils.deepClone(exampleData);
 
 const LegendHiddenGeneralExample = () => {
   React.useEffect(() => {
     const graph = Carbon.api.graph(graphConfig);
-    dataset.forEach((data) => {
-      graph.loadContent(Carbon.api.line(data));
-    });
+    graph.loadContent(Carbon.api.line(dataset));
   }, []);
   return (
     <ExampleGraphContainer id="legendHidden" />