+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs/General/Axes/SuppressTickValueTrailingZero.jsx
@@ -2,11 +2,12 @@ import React from 'react';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import LineGraph from '@cerner/terra-graphs-docs/lib/terra-graphs-src/components/Line/LineGraph';
 import '@cerner/terra-graphs-docs/lib/terra-dev-site/ExampleGraphContainer/ExampleGraphContainer.module.scss';
-import lineDecimalAxes from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/General/lineDecimalAxesConfig';
+import getGraphConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/General/generalDefaultDecimalAxes';
 
 /*
 Please refer to the documentation below to see the graphConfig and data objects
 */
-const graphConfig = utils.deepClone(lineDecimalAxes('#supressTickValue'));
+
+const graphConfig = utils.deepClone(getGraphConfig('#supressTickValue'));
 
 export default () => <LineGraph graphID="supressTickValue" graphConfig={graphConfig} />;