+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs/General/Grid/HorizontalGridHidden.jsx
@@ -2,13 +2,13 @@ import React from 'react';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import LineGraph from '@cerner/terra-graphs-docs/lib/terra-graphs-src/components/Line/LineGraph';
 import '@cerner/terra-graphs-docs/lib/terra-dev-site/ExampleGraphContainer/ExampleGraphContainer.module.scss';
-import lineDefault from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineDefault';
+import getGraphConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/General/generalDefaultHGridHidden';
 
 /*
 Please refer to the documentation below to see the graphConfig and data objects
 */
-const graphConfig = utils.deepClone(lineDefault('#horizontalGridHidden'));
-graphConfig.showHGrid = false;
+
+const graphConfig = utils.deepClone(getGraphConfig('#horizontalGridHidden'));
 
 export default () => <LineGraph graphID="horizontalGridHidden" graphConfig={graphConfig} />;