+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs/General/Other/BackgroundColor.jsx
@@ -2,14 +2,15 @@ import React from 'react';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import LineGraph from '@cerner/terra-graphs-docs/lib/terra-graphs-src/components/Line/LineGraph';
 import '@cerner/terra-graphs-docs/lib/terra-dev-site/ExampleGraphContainer/ExampleGraphContainer.module.scss';
-import lineDefault from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineDefault';
-import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/simpleLineData';
+import getGraphConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/General/generalDefault';
+import exampleData from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/General/dataset1';
 
 /*
 Please refer to the documentation below to see the graphConfig and data objects
 */
-const graphConfig = utils.deepClone(lineDefault('#backgroundColor'));
-const dataset = utils.deepClone(data);
+
+const graphConfig = utils.deepClone(getGraphConfig('#backgroundColor'));
+const dataset = [utils.deepClone(exampleData)];
 
 export default () => (
   <div className="background-color-class">