+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/General/Axes/DatetimeBucket.test.jsx
@@ -2,11 +2,11 @@ import React from 'react';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import LineGraph from '@cerner/terra-graphs-docs/lib/terra-graphs-src/components/Line/LineGraph';
 import '@cerner/terra-graphs-docs/lib/terra-dev-site/ExampleGraphContainer/ExampleGraphContainer.module.scss';
-import dateTimeBucket from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/General/dateTimeBucket';
-import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/timeseriesData';
+import getGraphConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/General/timeseriesConfigDateTimeBucket';
+import exampleData from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/General/datasetTimeseries1';
 
-const graphConfig = utils.deepClone(dateTimeBucket('#datetimeBucket'));
-const dataset = utils.deepClone(data);
+const graphConfig = utils.deepClone(getGraphConfig('#datetimeBucket'));
+const dataset = [utils.deepClone(exampleData)];
 
 export default () => (
   <>