+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/General/Axes/SuppressTickValueTrailingZero.test.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import LineGraph from '@cerner/terra-graphs-docs/lib/terra-graphs-src/components/Line/LineGraph';
 import '@cerner/terra-graphs-docs/lib/terra-dev-site/ExampleGraphContainer/ExampleGraphContainer.module.scss';
-import lineDecimalAxes from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/General/lineDecimalAxesConfig';
+import getGraphConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/General/generalDefaultDecimalAxes';
 
-const graphConfig = utils.deepClone(lineDecimalAxes('#supressTickValue'));
+const graphConfig = utils.deepClone(getGraphConfig('#supressTickValue'));
 
 export default () => <LineGraph graphID="supressTickValue" graphConfig={graphConfig} />;