+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/General/Axes/XAxisOrientationTop.test.jsx
@@ -1,11 +1,9 @@
 import React from 'react';
-import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import LineGraph from '@cerner/terra-graphs-docs/lib/terra-graphs-src/components/Line/LineGraph';
 import '@cerner/terra-graphs-docs/lib/terra-dev-site/ExampleGraphContainer/ExampleGraphContainer.module.scss';
-import lineDefault from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineDefault';
+import getGraphConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/General/generalDefaultXAxisTop';
 
-const graphConfig = utils.deepClone(lineDefault('#xAxisOrientationTop'));
-graphConfig.axis.x.orientation = Carbon.helpers.AXES_ORIENTATION.X.TOP;
+const graphConfig = utils.deepClone(getGraphConfig('#xAxisOrientationTop'));
 
 export default () => <LineGraph graphID="xAxisOrientationTop" graphConfig={graphConfig} />;