+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/General/Axes/XAxisTicksFormatted.test.jsx
@@ -2,20 +2,8 @@ import React from 'react';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import LineGraph from '@cerner/terra-graphs-docs/lib/terra-graphs-src/components/Line/LineGraph';
 import '@cerner/terra-graphs-docs/lib/terra-dev-site/ExampleGraphContainer/ExampleGraphContainer.module.scss';
-import lineTimeseries from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineTimeseries';
+import getGraphConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/General/timeseriesConfigTicksFormatted';
 
-const tickValues = [
-  new Date(2016, 0, 1, 1, 0).toISOString(),
-  new Date(2016, 0, 1, 5, 0).toISOString(),
-  new Date(2016, 0, 1, 10, 0).toISOString(),
-  new Date(2016, 0, 1, 15, 0).toISOString(),
-  new Date(2016, 0, 1, 20, 0).toISOString(),
-];
-
-const graphConfig = utils.deepClone(lineTimeseries('#xAxisTicksFormatted', 'LINE_TIMESERIES'));
-graphConfig.axis.x.ticks = {
-  values: tickValues,
-  format: '%a %b %e %X %Y',
-};
+const graphConfig = utils.deepClone(getGraphConfig('#xAxisTicksFormatted'));
 
 export default () => <LineGraph graphID="xAxisTicksFormatted" graphConfig={graphConfig} />;